Fix logger so passed-in config doesn't wipe all
Currently, if you pass-in some config parameters when the Logger is
instantiated, the parameters that you DON'T pass in do not default to
those in the config file.  This fix means the passed-in parameters only
override those included in the passed-in array, and others not specified
are taken from the config file.

Also change HttpError class to remove line which generated an error of
its own in the log.